# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of users"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/user/{id}/company/{companyId}": {
      "patch": {
        "operationId": "UsersController_setCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User company successfully updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set the company for a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/companies": {
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new company",
        "tags": [
          "Companies"
        ]
      },
      "get": {
        "operationId": "CompaniesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of companies with user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyWithUsersDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all companies with their users",
        "tags": [
          "Companies"
        ]
      }
    },
    "/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a company with user data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a company by ID with its users",
        "tags": [
          "Companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a company",
        "tags": [
          "Companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company successfully deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a company",
        "tags": [
          "Companies"
        ]
      }
    },
    "/companies/{id}/users/{userId}": {
      "post": {
        "operationId": "CompaniesController_addUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully added to company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a user to a company",
        "tags": [
          "Companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_removeUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully removed from company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyWithUsersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a user from a company",
        "tags": [
          "Companies"
        ]
      }
    },
...
  },
  "info": {
    "title": "Localize API",
    "description": "API documentation for Localize platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password (will be hashed)"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "owner",
              "manager",
              "member",
              "translator"
            ],
            "default": "member"
          },
          "isSystemAdmin": {
            "type": "boolean",
            "description": "Is user a system administrator",
            "default": false
          },
          "company": {
            "type": "string",
            "description": "ID of the company the user belongs to",
            "format": "ObjectId"
          }
        },
        "required": [
          "email",
          "password",
          "firstName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "avatarUrl": {
            "type": "string",
            "description": "User avatar URL"
          },
          "isSystemAdmin": {
            "type": "boolean",
            "description": "Is the user a system administrator"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "owner",
              "manager",
              "member",
              "translator"
            ]
          },
          "company": {
            "type": "string",
            "description": "User company",
            "format": "ObjectId"
          }
        }
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Company name"
          },
          "description": {
            "type": "string",
            "description": "Company description"
          },
          "projects": {
            "description": "Array of project IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name"
          },
          "avatarUrl": {
            "type": "string",
            "description": "User's avatar URL"
          },
          "isSystemAdmin": {
            "type": "boolean",
            "description": "Whether the user is a system admin",
            "default": false
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "owner",
              "manager",
              "member",
              "translator"
            ],
            "default": "member"
          },
          "company": {
            "type": "string",
            "description": "Company ID the user belongs to",
            "format": "ObjectId"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last login timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "isSystemAdmin",
          "role"
        ]
      },
      "CompanyWithUsersDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Company ID"
          },
          "name": {
            "type": "string",
            "description": "Company name"
          },
          "description": {
            "type": "string",
            "description": "Company description"
          },
          "projects": {
            "description": "Array of project IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          },
          "users": {
            "description": "Users belonging to this company",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Company name"
          },
          "description": {
            "type": "string",
            "description": "Company description"
          },
          "projects": {
            "description": "Array of project IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "ObjectId"
            }
          }
        }
      },
...
    }
  }
}